+++ CHANGES.md
@@ -6,9 +6,11 @@
   stratification, with fairness constraints enforced within each stratum, but
   not between strata
 * Update `make_derived_metric()` to use `MetricFrame`
+* Add `fairlearn.metrics_plots` module, with `plot_disparities_in_performance`
+  and `plot_disparities_in_metric` functions
 
 ### v0.5.0
-* Add `fairlearn.metrics_plots` module, with `plot_disparities_in_performance` and `plot_disparities_in_selection_rate` functions https://github.com/fairlearn/fairlearn/pull/561
+
 * Adjust classes to abide by naming conventions for attributes.
 * Change `ExponentiatedGradient` signature by renaming argument `T` to
   `max_iter`, `eta_mul` to `eta0`, and by adding `run_linprog_step`.